+++ packages/solana-wallet-snap/src/core/services/token-prices/types.ts
@@ -1,4 +1,0 @@
-import type { OnAssetHistoricalPriceResponse } from '@metamask/snaps-sdk';
-
-export type HistoricalPrice =
-  NonNullable<OnAssetHistoricalPriceResponse>['historicalPrice'];